requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
DateTime==5.5
dotenv==0.9.9
fastapi==0.121.2
fastapi-limiter==0.1.6
flatbuffers==25.9.23
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
h3==4.3.1
idna==3.11
numpy==2.3.4
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
sniffio==1.3.1
starlette==0.49.3
timezonefinder==8.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
zope.interface==8.1.1
